requirements.txt
ariadne==0.12.0
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
fastapi==0.61.1
graphql-core==3.0.5
h11==0.11.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pydantic==1.7.2
pylint==2.6.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.12.2
websockets==8.1
wrapt==1.12.1
